refactor: automatically add identity proof
